+++ Domain/IEntity/IEntity.cs
@@ -3,11 +3,11 @@ namespace Domain.IEntity
 {
     public interface IEntity
     {
-        int Id { get; set; }
+        long Id { get; set; }
     }
 
     public class Entity: IEntity
     {
-       public virtual int Id { get; set; }
+       public virtual long Id { get; set; }
     }
 }
